task_mAIstro/task_mAIstro/requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
trustcall
ipython

# Gmail integration dependencies
requests
httpx
pydantic
python-dotenv
fastapi
uvicorn
python-multipart

# Voice interface dependencies
openai
elevenlabs
numpy
sounddevice
soundfile